+++ src/src_lyrics.h
@@ -71,7 +71,7 @@ typedef int (*src_lyr_plugin_register) (src_lyr *source_descriptor);
 
 GArray *src_lyr_stack;
 
-int get_text_line(formed_text *text, int num, char *dest, int len);
+int get_text_line(formed_text *text, unsigned num, char *dest, size_t len);
 
 void src_lyr_stack_init(void);
 int src_lyr_init(void);
